pruning start
