+++ src/pyhf/cli/__init__.py
@@ -1,0 +1,5 @@
+from .rootio import cli as rootio
+from .spec import cli as spec
+from .stats import cli as stats
+
+__all__ = ['rootio', 'spec', 'stats']
